Transition to Lua
There are a lot of improvements to the speed, but I think it's just a
natural limitation of how Vim stores highlighting information.
Unfortunately, we can't really get at that information very quickly...
however, if you're moving to things on the same line or close to your
line, the delay is slight but not painful.

Try not to search several thousand lines.... :)

Could probably put a timeout on it or something.